src/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
pymysql==1.1.0         # Use this if you're using MySQL
# motor==3.4.0          # Uncomment if you're using MongoDB (FastAPI async with Mongo)

# Pydantic v2 with .env support
pydantic==2.11.1
pydantic-settings==2.2.1

# Alembic for migrations (optional)
alembic==1.13.1

# Authentication / JWT
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Email, date handling, etc.
python-multipart==0.0.9
email-validator==2.1.1
python-dotenv==1.0.1

# Celery (background tasks) and Redis (broker)
celery==5.3.6
redis==5.0.4

# Optional: linting & formatting
black==24.3.0
isort==5.13.2
flake8==7.0.0